feat: 해당 테마에 대한 예약이 있을 경우 삭제 시 예외 구현

+++ src/main/java/roomescape/reservation/service/ThemeService.java
@@ -5,15 +5,18 @@ import org.springframework.stereotype.Service;
 import roomescape.reservation.controller.dto.ThemeRequest;
 import roomescape.reservation.controller.dto.ThemeResponse;
 import roomescape.reservation.domain.Theme;
+import roomescape.reservation.domain.repository.ReservationRepository;
 import roomescape.reservation.domain.repository.ThemeRepository;
 
 @Service
 public class ThemeService {
 
     private final ThemeRepository themeRepository;
+    private final ReservationRepository reservationRepository;
 
-    public ThemeService(ThemeRepository themeRepository) {
+    public ThemeService(ThemeRepository themeRepository, ReservationRepository reservationRepository) {
         this.themeRepository = themeRepository;
+        this.reservationRepository = reservationRepository;
     }
 
     public ThemeResponse add(ThemeRequest request) {
@@ -31,6 +34,10 @@ public class ThemeService {
     }
 
     public void remove(Long id) {
-        themeRepository.deleteById(id);
+        if (!reservationRepository.existReservationByThemeId(id)) {
+            themeRepository.deleteById(id);
+            return;
+        }
+        throw new IllegalStateException("해당 테마에 대한 예약이 존재합니다.");
     }
 }
